Stage 1 - new auth
Added requests based auth using api_token. This is still not working with username/password but it's a valid auth for the old APIs so any other method will still work. Added _reqs / a _req copy using requests instead of `httplib2.Http` and the first method using it (search_user).